Move jobs out of views.py
