+++ core/views/__init__.py
@@ -2,3 +2,4 @@ from .views import *
 from .validation_scenario import *
 from .organization import *
 from .record_group import *
+from .job import *
